requirements.txt
# Azure DevOps Governance Factory - Enhanced Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure SDK and Integration
azure-devops==7.1.0b4
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-monitor-opentelemetry==1.1.0
azure-mgmt-core==1.4.0

# Our comprehensive Azure DevOps wrapper dependencies
asyncio-throttle==1.0.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=42.0.0

# Database and Caching
sqlalchemy>=2.0.0
alembic>=1.13.0
redis<5.0.0,>=4.5.2
aioredis>=2.0.0
psycopg2-binary>=2.9.0

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Task Queue and Background Jobs
celery[redis]==5.3.4
kombu==5.3.4
flower==2.0.1

# Machine Learning and Analytics
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Configuration and Environment
python-dotenv==1.0.0
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# API Documentation
swagger-ui-bundle==0.0.9
redoc==1.0.1

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
docker==6.1.3

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-swagger-ui-tag==0.6.8
